+++ addOns/customactivescan/src/main/java/org/zaproxy/zap/extension/customactivescan/HttpMessageWithLCSResponse.java
@@ -1,23 +1,30 @@
 package org.zaproxy.zap.extension.customactivescan;
 
 import org.parosproxy.paros.core.scanner.Alert;
+import org.parosproxy.paros.network.HttpMalformedHeaderException;
 import org.parosproxy.paros.network.HttpMessage;
 import org.zaproxy.zap.extension.customactivescan.model.AttackTitleType;
+import org.zaproxy.zap.extension.customactivescan.view.InterfacePopUpAction;
 
+import java.awt.*;
 import java.util.List;
 
-public class HttpMessageWithLCSResponse extends HttpMessage {
-    private String lcsResponse;
-    private int originalAverageResponseSize;
-    private int worstResponseStatus;
+public class HttpMessageWithLCSResponse extends HttpMessage implements InterfacePopUpAction {
+    private final static org.apache.logging.log4j.Logger LOGGER4J =
+            org.apache.logging.log4j.LogManager.getLogger();
+    private String lcsResponse = null;
+    private int originalAverageResponseSize = -1;
+    private int worstResponseStatus = -1;
     private int messageIndexInScanLogPanel; //  index of message in ScanLogPanel.resultMessageList
     private List<StartEndPosition> lcsCharacterIndexOfLcsResponse = null;
     private List<StartEndPosition> lcsCharacterIndexOfLcsRequest = null;
-    private String attackTitleString;
+    private String attackTitleString = null;
     private String percentString = "";
     private int injectionPatternGroupIndex = -1;
     private Alert alert = null;
-    private AttackTitleType attackTitleType;
+    private AttackTitleType attackTitleType = null;
+    private Component popUpInvokerComponent = null;
+    private ParmGenMacroTraceParams pmtParams = null;
 
     HttpMessageWithLCSResponse(
             HttpMessage htmsg,
@@ -44,6 +51,26 @@ public class HttpMessageWithLCSResponse extends HttpMessage {
         this.originalAverageResponseSize = entireResponseString.length();
         this.attackTitleType = attackTitleType;
         this.attackTitleString = attackTitleType.name();
+    }
+
+    private HttpMessageWithLCSResponse(HttpMessageWithLCSResponse message) {
+        super(message);
+    }
+
+    private void copyMembersToDest(HttpMessageWithLCSResponse dest) {
+        dest.lcsResponse = this.lcsResponse;
+        dest.originalAverageResponseSize = this.originalAverageResponseSize;
+        dest.worstResponseStatus = this.worstResponseStatus;
+        dest.messageIndexInScanLogPanel = this.messageIndexInScanLogPanel;
+        dest.lcsCharacterIndexOfLcsResponse = this.lcsCharacterIndexOfLcsResponse;
+        dest.lcsCharacterIndexOfLcsRequest = this.lcsCharacterIndexOfLcsRequest;
+        dest.attackTitleString = this.attackTitleString;
+        dest.percentString = this.percentString;
+        dest.injectionPatternGroupIndex = this.injectionPatternGroupIndex;
+        dest.alert = this.alert;
+        dest.attackTitleType = this.attackTitleType;
+        dest.popUpInvokerComponent = this.popUpInvokerComponent;
+        dest.pmtParams = this.pmtParams;
     }
 
     /**
@@ -133,4 +160,49 @@ public class HttpMessageWithLCSResponse extends HttpMessage {
     public Alert getAlert() {
         return this.alert;
     }
+
+    public void setPopUpInvokerComponent(Component component) {
+        this.popUpInvokerComponent = component;
+    }
+    @Override
+    public void popUpActionPerformed(HttpMessage message) {
+        if (this.popUpInvokerComponent != null) {
+            if (InterfacePopUpAction.class.isAssignableFrom(this.popUpInvokerComponent.getClass())){
+                InterfacePopUpAction action = (InterfacePopUpAction) this.popUpInvokerComponent;
+                action.popUpActionPerformed(message);
+            }
+        }
+    }
+
+
+    @Override
+    public HttpMessageWithLCSResponse cloneRequest() {
+        HttpMessageWithLCSResponse newMessage = new HttpMessageWithLCSResponse(this);
+        if (newMessage != null) {
+            copyMembersToDest(newMessage);
+        }
+        return newMessage;
+    }
+
+    public void hello(String data, Integer i) {
+
+        LOGGER4J.info("hello from customactive data:" + data + " i=" + i.toString());
+    }
+
+    public void setPmtParams(ParmGenMacroTraceParams pmtParams) {
+        this.pmtParams = pmtParams;
+    }
+
+    public Integer[] getPmtParamsArray() {
+        if (this.pmtParams != null) {
+            return new Integer[] {
+                    this.pmtParams.scannerId,
+                    this.pmtParams.getSelectedRequestNo(),
+                    this.pmtParams.getLastStepNo(),
+                    this.pmtParams.getTabIndex()
+            };
+        }
+        return null;
+    }
+
 }