+++ addOns/customactivescan/src/main/java/org/zaproxy/zap/extension/customactivescan/view/InterfacePopUpAction.java
@@ -1,0 +1,7 @@
+package org.zaproxy.zap.extension.customactivescan.view;
+
+import org.parosproxy.paros.network.HttpMessage;
+
+public interface InterfacePopUpAction {
+    public void popUpActionPerformed(HttpMessage message);
+}